Task 4 completed; small fixes
