+++ lab-2/src/main/java/by/bsu/dependency/Exception/NotSingletonException.java
@@ -1,0 +1,7 @@
+package by.bsu.dependency.Exception;
+
+public class NotSingletonException extends RuntimeException {
+    public NotSingletonException(String message) {
+        super(message);
+    }
+}
