+++ lab-2/src/main/java/by/bsu/dependency/context/AutoScanApplicationContext.java
@@ -1,5 +1,11 @@
 package by.bsu.dependency.context;
 
+import by.bsu.dependency.annotation.Bean;
+import org.reflections.Reflections;
+import org.reflections.scanners.SubTypesScanner;
+
+import java.util.List;
+
 public class AutoScanApplicationContext extends AbstractApplicationContext {
 
     /**
@@ -12,51 +18,14 @@ public class AutoScanApplicationContext extends AbstractApplicationContext {
      * @param packageName имя сканируемого пакета
      */
     public AutoScanApplicationContext(String packageName) {
-        throw new IllegalStateException("not implemented");
+        super(findAllBeanClasses(packageName));
     }
 
-    @Override
-    public void start() {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public boolean isRunning() {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public boolean containsBean(String name) {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public boolean containsBean(Class<?> clazz) {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public boolean containsBeanDefinition(String name) {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public Object getBean(String name) {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public <T> T getBean(Class<T> clazz) {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public boolean isPrototype(String name) {
-        throw new IllegalStateException("not implemented");
-    }
-
-    @Override
-    public boolean isSingleton(String name) {
-        throw new IllegalStateException("not implemented");
+    private static List<Class<?>> findAllBeanClasses(String packageName) {
+        Reflections reflections = new Reflections(packageName, new SubTypesScanner(false));
+        return reflections.getSubTypesOf(Object.class)
+                .stream()
+                .filter(clazz -> clazz.isAnnotationPresent(Bean.class))
+                .toList();
     }
 }
