+++ lab-2/src/main/java/by/bsu/dependency/example/ContainsNotIncludedBean.java
@@ -4,8 +4,8 @@ import by.bsu.dependency.annotation.Bean;
 import by.bsu.dependency.annotation.BeanScope;
 import by.bsu.dependency.annotation.Inject;
 
-@Bean(name="containesNotIncludedBean", scope= BeanScope.PROTOTYPE)
-public class ContainesNotIncludedBean {
+@Bean(name="containsNotIncludedBean", scope= BeanScope.PROTOTYPE)
+public class ContainsNotIncludedBean {
 
     @Inject
     NotIncludedBean notIncludedBean;
